+++ ios/CameraeVision/include/CameraeVision.h
@@ -3,4 +3,5 @@
 #import <CameraeVision/CameraeVisionRuntime.h>
 #import <CameraeVision/CameraeVisionCaptureSession.h>
 #import <CameraeVision/CameraeVisionClipAlignment.h>
+#import <CameraeVision/CameraeVisionPlateSolver.h>
 #import <CameraeVision/OpenCVBridge.h>